reference/ethchiangmai-hackathon-2026/services/ai/requirements.txt
# Web 框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# AI 模型
funasr>=1.1.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
modelscope>=1.10.0
scipy

# BLS 签名
py-ecc==6.0.0

# 工具
python-dotenv==1.0.0
